Remove Rager releases badge
I'm the maintainer of the Rager project. I haven't had much time to dedicate to the project lately, so it hasn't received a lot of love. The site has gone down a few times lately, which caused requests for the badge to fail, and I wasn't able to quickly resolve the problem. I suggest that you remove the badge from your README, since a failed request for the badge can make the front page of your repo look a little messy. I'm sorry for the failures in the past.

For those looking for a similar project with better support/features, you should check out Libraries.io. I have no affiliation with that project, but it seems to tackle some of the same problems I wanted to address with Rager.